+++ packages/terra-form-field/src/Field.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { injectIntl, intlShape } from 'react-intl';
 import classNames from 'classnames/bind';
 import IconError from 'terra-icon/lib/icon/IconError';
 import styles from './Field.module.scss';
@@ -32,11 +31,6 @@ const propTypes = {
    * The htmlFor attribute on the field label.
    */
   htmlFor: PropTypes.string,
-  /**
-   * @private
-   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
-   */
-  intl: intlShape.isRequired,
   /**
    * Whether or not the field is invalid.
    */
@@ -94,12 +88,21 @@ const defaultProps = {
   showOptional: false,
 };
 
+const contextTypes = {
+  /* eslint-disable consistent-return */
+  intl: (context) => {
+    if (context.intl === undefined) {
+      return new Error('Component is internationalized, and must be wrapped in terra-base');
+    }
+  },
+};
+
 const hasWhiteSpace = s => /\s/g.test(s);
 // Detect IE 10 or IE 11
 // TODO - Delete detection logic when we drop support for IE
 const isIE = () => (window.navigator.userAgent.indexOf('Trident/6.0') > -1 || window.navigator.userAgent.indexOf('Trident/7.0') > -1);
 
-const Field = (props) => {
+const Field = (props, { intl }) => {
   const {
     children,
     error,
@@ -107,7 +110,6 @@ const Field = (props) => {
     help,
     hideRequired,
     htmlFor,
-    intl,
     isInvalid,
     isInline,
     isLabelHidden,
@@ -201,5 +203,6 @@ const Field = (props) => {
 
 Field.propTypes = propTypes;
 Field.defaultProps = defaultProps;
+Field.contextTypes = contextTypes;
 
-export default injectIntl(Field);
+export default Field;