+++ packages/terra-form-select/tests/jest/MaxSelection.test.jsx
@@ -5,7 +5,7 @@ import MaxSelection from '../../src/_MaxSelection';
 
 describe('MaxSelection', () => {
   it('should render a default MaxSelection', () => {
-    const wrapper = shallowWithIntl(<MaxSelection value="3" />).dive();
+    const wrapper = shallowWithIntl(<MaxSelection value="3" />);
     expect(wrapper).toMatchSnapshot();
   });
 });