+++ packages/terra-form-select/tests/jest/NoResults.test.jsx
@@ -5,12 +5,12 @@ import NoResults from '../../src/_NoResults';
 
 describe('NoResults', () => {
   it('should render a default NoResults', () => {
-    const wrapper = shallowWithIntl(<NoResults value="value" />).dive();
+    const wrapper = shallowWithIntl(<NoResults value="value" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render NoResults with custom content', () => {
-    const wrapper = shallowWithIntl(<NoResults value="value" noResultContent={<div>Custom</div>} />).dive();
+    const wrapper = shallowWithIntl(<NoResults value="value" noResultContent={<div>Custom</div>} />);
     expect(wrapper).toMatchSnapshot();
   });
 });