+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Updated
 * updated jest snapshots
 
+3.19.0 - (July 17, 2019)
+------------------
+### Changed
+* Revert components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+
 3.18.0 - (July 17, 2019)
 ------------------
 ### Changed
@@ -19,7 +24,6 @@ Unreleased
 3.16.0  - (July 11, 2019)
 ------------------
 ### Changed
-* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports
 * Updated Docs around the requirement of terra-base and aggregate-translations.